requirements.txt
alembic==1.13.3
amqp==5.3.1
annotated-types==0.7.0
billiard==4.2.1
blinker==1.8.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
CryptoEngine @ git+https://github.com/CryptoToolsLtd/CryptoEngine.git@d5cca595b6ad12ccc71d539d25c5026e346822e1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
Mako==1.3.6
MarkupSafe==2.1.5
mysqlclient==2.2.5
prompt_toolkit==3.0.48
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
